+++ backend/src/main/java/com/optimize/elykia/core/service/accounting/AccountingDayService.java
@@ -5,6 +5,7 @@ import com.optimize.common.entities.exception.ResourceNotFoundException;
 import com.optimize.common.entities.service.GenericService;
 import com.optimize.elykia.core.dto.CloseCollectorOperationDto;
 import com.optimize.elykia.core.entity.accounting.AccountingDay;
+import com.optimize.elykia.core.entity.accounting.DailyAccountancy;
 import com.optimize.elykia.core.entity.accounting.DailyAccounting;
 import com.optimize.elykia.core.enumaration.AccountingDayStatus;
 import com.optimize.elykia.core.repository.AccountingDayRepository;